+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## MASTER
+## Dotenv 0.5.3.0
+* Move `loadSafeFile` to `Configuration.Dotenv`
+* Change `loadSafeFile` signature
+* Add `type ValidatorMap = Map Text (Text -> Bool)` for custom validations.
 
 ## Dotenv 0.5.2.5
 * Update `exceptions` bounds `>= 0.8 && < 0.11`